requirements.txt
discord.py[voice]==2.3.2
aiohttp==3.9.3
pynacl==1.5.0